Enable marking elements for removal
Adds a removeElements option taking a callback function to indicate that
the element should be marked for removal. The existing attributes will
be replaced with a class consisting of the removal class name, also set
in options (removalClassName).

This enables mobileapps' rmElements transform to search for only a
single class name, rather than several more-complex selectors, when
removing unwanted elements.